Rename JavaJson to OrgAndrejsJson
