+++ Samples/JavaDependencySampleApp/Sources/OrgAndrejsJson/dummy.swift
