+++ Samples/JavaDependencySampleApp/Sources/OrgAndrejsJson/swift-java.config
